아이템 습득(보스 유실물 및 자연 물품) 클래스,  event, UI 추가/ 아이템 물품 조금 추가함
