+++ Assets/Game/Scripts/Item/ItemInterAction(YJ)/CollectItemUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb9fb87cd8e97b24dbe4baba713d66c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
