+++ Assets/Game/Scripts/Item/ItemInterAction(YJ)/InteractionNature.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ec5dc95206c3e642803f52c6ca9f4e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
