--- Assets/Game/Scripts/Item/ItemInterAction(YJ)/InteractionNatureObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f63075679f013d64d94f56372539af48
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
